requirements.txt
Django==4.2.9
djangorestframework==3.15.2
PyJwt==2.3.0
django-cors-headers==4.4.0